+++ docker/gunicorn.sh
@@ -8,4 +8,4 @@ exec gunicorn \
   -w 5 \
   -b 0.0.0.0:8000 \
   --timeout 120 \
-  treeherder.webapp.wsgi:application
+  treeherder.config.wsgi:application
