+++ treeherder/settings/base.py
@@ -33,7 +33,7 @@ SECRET_KEY = os.environ.get("TREEHERDER_DJANGO_SECRET_KEY")
 
 SITE_ID = 1
 ROOT_URLCONF = "treeherder.webapp.urls"
-WSGI_APPLICATION = 'treeherder.webapp.wsgi.application'
+WSGI_APPLICATION = 'treeherder.config.wsgi.application'
 
 TIME_ZONE = "America/Los_Angeles"
 USE_I18N = False
